fixed a few typo bugs, added helper method to get current frequency

+++ src/main/java/com/jbatista/wmo/synthesis/EnvelopeGenerator.java
@@ -101,7 +101,7 @@ public class EnvelopeGenerator {
         this.factor[1] = 1d / size[1];
     }
 
-    public double getSustainSpeed() {
+    public int getSustainSpeed() {
         return sustainSpeed;
     }
